Add files via upload
Control flow in Python:  order in which statements are executed. Decision control uses if, elif, and else statements to execute specific blocks of code based on conditions. These statements evaluate boolean expressions (True or False) and direct program flow accordingly. Indentation in Python is critical to defining blocks of code for conditional statements.